Add chunks to voxelgrid generation

+++ VulkanEngine/GameObject.h
@@ -4,7 +4,7 @@
 
 struct GameObject
 {
-	uint8_t gameObjectId = 0, textureId = 0, modelId = 0;
+	size_t gameObjectId = 0, textureId = 0, modelId = 0;
 
 	glm::vec3 position, rotation;
